[REF] spreadsheet(_account): remove useless data sources
The currency data source and accounting data source are actually quite
useless. They do not have any state (maybe they used to have the cache).

The respective plugins were only acting as proxies between the functions
implementation and the data sources. There's no value to have the two
different things.
Everything can be moved to the plugins. There's less code, less useless plumbing
that needs to be maintained.
One could even argue that we could have everything in the `compute` by
exposing `serverData`.

Task: 3736186
