+++ addons/spreadsheet_account/static/src/plugins/accounting_plugin.js
@@ -1,8 +1,10 @@
 /** @odoo-module */
+// @ts-check
 
-import { AccountingDataSource } from "../accounting_datasource";
+import { EvaluationError } from "@odoo/o-spreadsheet";
 import { OdooUIPlugin } from "@spreadsheet/plugins";
-const DATA_SOURCE_ID = "ACCOUNTING_AGGREGATES";
+import { _t } from "@web/core/l10n/translation";
+import { camelToSnakeObject, toServerDateString } from "@spreadsheet/helpers/helpers";
 
 /**
  * @typedef {import("../accounting_functions").DateRange} DateRange
@@ -18,10 +20,17 @@ export class AccountingPlugin extends OdooUIPlugin {
     ]);
     constructor(config) {
         super(config);
-        this.dataSources = config.custom.dataSources;
-        if (this.dataSources) {
-            this.dataSources.add(DATA_SOURCE_ID, AccountingDataSource);
+        /** @type {import("@spreadsheet/data_sources/server_data").ServerData} */
+        this._serverData = config.custom.dataSources?.serverData;
+    }
+
+    get serverData() {
+        if (!this._serverData) {
+            throw new Error(
+                "'serverData' is not defined, please make sure a 'DataSources' instance is provided to the model."
+            );
         }
+        return this._serverData;
     }
 
     // -------------------------------------------------------------------------
@@ -38,12 +47,8 @@ export class AccountingPlugin extends OdooUIPlugin {
      * @returns {number}
      */
     getAccountPrefixCredit(codes, dateRange, offset, companyId, includeUnposted) {
-        return (
-            this.dataSources &&
-            this.dataSources
-                .get(DATA_SOURCE_ID)
-                .getCredit(codes, dateRange, offset, companyId, includeUnposted)
-        );
+        const data = this._fetchAccountData(codes, dateRange, offset, companyId, includeUnposted);
+        return data.credit;
     }
 
     /**
@@ -56,12 +61,8 @@ export class AccountingPlugin extends OdooUIPlugin {
      * @returns {number}
      */
     getAccountPrefixDebit(codes, dateRange, offset, companyId, includeUnposted) {
-        return (
-            this.dataSources &&
-            this.dataSources
-                .get(DATA_SOURCE_ID)
-                .getDebit(codes, dateRange, offset, companyId, includeUnposted)
-        );
+        const data = this._fetchAccountData(codes, dateRange, offset, companyId, includeUnposted);
+        return data.debit;
     }
 
     /**
@@ -70,10 +71,7 @@ export class AccountingPlugin extends OdooUIPlugin {
      * @returns {string | undefined}
      */
     getFiscalStartDate(date, companyId) {
-        return (
-            this.dataSources &&
-            this.dataSources.get(DATA_SOURCE_ID).getFiscalStartDate(date, companyId)
-        );
+        return this._fetchCompanyData(date, companyId).start;
     }
 
     /**
@@ -82,10 +80,7 @@ export class AccountingPlugin extends OdooUIPlugin {
      * @returns {string | undefined}
      */
     getFiscalEndDate(date, companyId) {
-        return (
-            this.dataSources &&
-            this.dataSources.get(DATA_SOURCE_ID).getFiscalEndDate(date, companyId)
-        );
+        return this._fetchCompanyData(date, companyId).end;
     }
 
     /**
@@ -93,9 +88,51 @@ export class AccountingPlugin extends OdooUIPlugin {
      * @returns {string[]}
      */
     getAccountGroupCodes(accountType) {
-        return (
-            this.dataSources &&
-            this.dataSources.get(DATA_SOURCE_ID).getAccountGroupCodes(accountType)
+        return this.serverData.batch.get("account.account", "get_account_group", accountType);
+    }
+
+    /**
+     * Fetch the account information (credit/debit) for a given account code
+     * @private
+     * @param {string[]} codes prefix of the accounts' codes
+     * @param {DateRange} dateRange start date of the period to look
+     * @param {number} offset end  date of the period to look
+     * @param {number | null} companyId specific companyId to target
+     * @param {boolean} includeUnposted wether or not select unposted entries
+     * @returns {{ debit: number, credit: number }}
+     */
+    _fetchAccountData(codes, dateRange, offset, companyId, includeUnposted) {
+        dateRange.year += offset;
+        // Excel dates start at 1899-12-30, we should not support date ranges
+        // that do not cover dates prior to it.
+        // Unfortunately, this check needs to be done right before the server
+        // call as a date to low (year <= 1) can raise an error server side.
+        if (dateRange.year < 1900) {
+            throw new EvaluationError(_t("%s is not a valid year.", dateRange.year));
+        }
+        return this.serverData.batch.get(
+            "account.account",
+            "spreadsheet_fetch_debit_credit",
+            camelToSnakeObject({ dateRange, codes, companyId, includeUnposted })
         );
+    }
+
+    /**
+     * Fetch the start and end date of the fiscal year enclosing a given date
+     * Defaults on the current user company if not provided
+     * @private
+     * @param {Date} date
+     * @param {number | null} companyId
+     * @returns {{start: string, end: string}}
+     */
+    _fetchCompanyData(date, companyId) {
+        const result = this.serverData.batch.get("res.company", "get_fiscal_dates", {
+            date: toServerDateString(date),
+            company_id: companyId,
+        });
+        if (result === false) {
+            throw new EvaluationError(_t("The company fiscal year could not be found."));
+        }
+        return result;
     }
 }
